+++ nilearn/plotting/html_surface.py
@@ -4,11 +4,14 @@ import collections.abc
 import numpy as np
 import matplotlib as mpl
 from matplotlib import cm as mpl_cm
+from functools import partial
 
 from nilearn._utils.niimg_conversions import check_niimg_3d
 from nilearn._utils import fill_doc
+from nilearn._utils.helpers import rename_parameters
 from nilearn import surface
 from nilearn import datasets
+from nilearn.plotting.surf_plotting import _deprecate_separate_mesh_data
 from nilearn.plotting.html_document import HTMLDocument
 from nilearn.plotting import cm
 from nilearn.plotting.js_plotting_utils import (
@@ -21,15 +24,19 @@ class SurfaceView(HTMLDocument):
 
 
 def _get_vertexcolor(surf_map, cmap, norm,
-                     absolute_threshold=None, bg_map=None):
+                     absolute_threshold=None, bg_surf=None):
     vertexcolor = cmap(norm(surf_map).data)
     if absolute_threshold is None:
         return to_color_strings(vertexcolor)
-    if bg_map is None:
+    if bg_surf is None:
         bg_map = np.ones(len(surf_map)) * .5
         bg_vmin, bg_vmax = 0, 1
     else:
-        bg_map = surface.load_surf_data(bg_map)
+        if(hasattr(bg_surf, 'mesh')
+           and hasattr(bg_surf, 'data')):
+            bg_map = bg_surf.data
+        else:
+            bg_map = surface.load_surf_data(bg_surf)
         bg_vmin, bg_vmax = np.min(bg_map), np.max(bg_map)
     bg_norm = mpl.colors.Normalize(vmin=bg_vmin, vmax=bg_vmax)
     bg_color = mpl_cm.get_cmap('Greys')(bg_norm(bg_map))
@@ -38,9 +45,9 @@ def _get_vertexcolor(surf_map, cmap, norm,
     return to_color_strings(vertexcolor)
 
 
-def one_mesh_info(surf_map, surf_mesh, threshold=None, cmap=cm.cold_hot,
-                  black_bg=False, bg_map=None, symmetric_cmap=True,
-                  vmax=None, vmin=None):
+def _one_mesh_info(surf_map, surf_mesh, threshold=None, cmap=cm.cold_hot,
+                   black_bg=False, bg_surf=None, symmetric_cmap=True,
+                   vmax=None, vmin=None):
     """Prepare info for plotting one surface map on a single mesh.
 
     This computes the dictionary that gets inserted in the web page,
@@ -55,7 +62,7 @@ def one_mesh_info(surf_map, surf_mesh, threshold=None, cmap=cm.cold_hot,
     info['inflated_left'] = mesh_to_plotly(surf_mesh)
     info['vertexcolor_left'] = _get_vertexcolor(
         surf_map, colors['cmap'], colors['norm'],
-        colors['abs_threshold'], bg_map)
+        colors['abs_threshold'], bg_surf)
     info["cmin"], info["cmax"] = float(colors['vmin']), float(colors['vmax'])
     info['black_bg'] = black_bg
     info['full_brain_mesh'] = False
@@ -230,7 +237,13 @@ def view_img_on_surf(stat_map_img, surf_mesh='fsaverage5',
     return _fill_html_template(info, embed_js=True)
 
 
-def view_surf(surf_mesh, surf_map=None, bg_map=None, threshold=None,
+deprecate_separate_mesh_data_view_surf = partial(
+    _deprecate_separate_mesh_data, argument="surf_map")
+
+
+@rename_parameters({'bg_map': 'bg_surf'}, '0.10.0')
+@deprecate_separate_mesh_data_view_surf
+def view_surf(surf_mesh, surf_map=None, *, bg_surf=None, threshold=None,
               cmap=cm.cold_hot, black_bg=False, vmax=None, vmin=None,
               symmetric_cmap=True, colorbar=True, colorbar_height=.5,
               colorbar_fontsize=25, title=None, title_fontsize=25):
@@ -238,13 +251,41 @@ def view_surf(surf_mesh, surf_map=None, bg_map=None, threshold=None,
 
     Parameters
     ----------
-    surf_mesh : str or list of two numpy.ndarray
-        Surface mesh geometry, can be a file (valid formats are
-        .gii or Freesurfer specific files such as .orig, .pial,
-        .sphere, .white, .inflated) or
-        a list of two Numpy arrays, the first containing the x-y-z coordinates
-        of the mesh vertices, the second containing the indices
-        (into coords) of the mesh faces.
+    surf_mesh : str or list of two numpy.ndarray or Mesh or Surface
+
+            .. deprecated:: 0.8.1
+                `surf_mesh` is deprecated in 0.8.1 and will be renamed
+                'surface' in 0.10.0.
+
+        If a Surface-like object is provided instead of a Mesh, `surface.data`
+        will overwrite the `surf_map` argument.
+        It can be:
+        - a Surface-like object with a mesh and data attributes
+
+        A surface can be:
+            - a nilearn.surface.Surface
+            - a sequence (mesh, data) where:
+
+            A mesh can be:
+                - a nilearn.surface.Mesh
+                - a path to .gii or .gii.gz etc.
+                - a sequence of two numpy arrays,
+                  the first containing vertex coordinates
+                  and the second containing triangles.
+
+            Data can be:
+                - a path to .gii or .gii.gz etc.
+                - a numpy array with shape (n_vertices,)
+                  or (n_time_point, n_vertices)
+
+        - a surface mesh geometry (deprecated)
+
+        A Mesh can be:
+            - a file (valid formats are .gii or Freesurfer specific
+              files such as .orig, .pial, .sphere, .white, .inflated)
+            - a list of two Numpy arrays, the first containing the
+              x-y-z coordinates of the mesh vertices, the second
+              containing the indices (into coords) of the mesh faces.
 
     surf_map : str or numpy.ndarray, optional
         Data to be displayed on the surface mesh. Can be a file (valid formats
@@ -252,10 +293,20 @@ def view_surf(surf_mesh, surf_map=None, bg_map=None, threshold=None,
         .thickness, .area, .curv, .sulc, .annot, .label) or
         a Numpy array
 
-    bg_map : Surface data, optional
-        Background image to be plotted on the mesh underneath the
-        surf_data in greyscale, most likely a sulcal depth map for
-        realistic shading.
+            .. deprecated:: 0.8.1
+                `surf_map` is deprecated in 0.8.1 and will be
+                removed in 0.10.0. Please use a Surface object
+                to define the map. This will not be used if a
+                Surface is provided as first argument.
+
+    bg_surf : Surface, optional
+        Background surface to be plotted on `surface.mesh` underneath
+        `surface.data` in greyscale. `bg_surf.data` is most likely a
+        sulcal depth map for realistic shading.
+
+        .. versionchanged:: 0.8.1
+            `bg_surf` was introduced in 0.8.1 and replaces `bg_map`.
+            `bg_map` will not be supported after release 0.10.0.
 
     threshold : str, number or None, optional
         If None, no thresholding.
@@ -324,11 +375,13 @@ def view_surf(surf_mesh, surf_map=None, bg_map=None, threshold=None,
     else:
         surf_mesh, surf_map = surface.check_mesh_and_data(
             surf_mesh, surf_map)
-    if bg_map is not None:
-        _, bg_map = surface.check_mesh_and_data(surf_mesh, bg_map)
-    info = one_mesh_info(
+    if bg_surf is not None:
+        if(not hasattr(bg_surf, 'mesh')
+           or not hasattr(bg_surf, 'data')):
+            _, bg_surf = surface.check_mesh_and_data(surf_mesh, bg_surf)
+    info = _one_mesh_info(
         surf_map=surf_map, surf_mesh=surf_mesh, threshold=threshold,
-        cmap=cmap, black_bg=black_bg, bg_map=bg_map,
+        cmap=cmap, black_bg=black_bg, bg_surf=bg_surf,
         symmetric_cmap=symmetric_cmap, vmax=vmax, vmin=vmin)
     info['colorbar'] = colorbar
     info['cbar_height'] = colorbar_height